move startMockServer import to mod
