+++ dev_utils/mod.ts
@@ -1,3 +1,4 @@
 export { createMockQueueSockConn, createMockSockConn } from "./mockConn.ts";
 export { createWebStreamPair } from "./webStreamPair.ts";
 export { resolveAsap, resolveNextTick, runAsap } from "./timers.ts";
+export { startMockServer } from "./mockServer.ts";
